refactor(token-select-hook-form): styles
