+++ features/wsteth/wrap/wrap-form-controls/token-select-wrap.tsx
@@ -2,7 +2,7 @@ import { trackEvent } from '@lidofinance/analytics-matomo';
 
 import { TOKENS_TO_WRAP } from 'features/wsteth/shared/types';
 import { MATOMO_CLICK_EVENTS } from 'consts/matomo-click-events';
-import { TokenSelectHookForm } from 'shared/hook-form/controls/token-select-hook-form';
+import { TokenSelectHookForm } from 'shared/hook-form/controls/token-select-hook-form/token-select-hook-form';
 import { useDappStatus } from 'shared/hooks/use-dapp-status';
 
 const OPTIONS = [
